Inline, Internal and External CSS
There are three ways to add CSS in HTML Page . 1) Inline CSS in which we use style attributes. 2)Internal CSS in which we use style Tag. 3) External CSS in which we create another CSS file and give Tags properties and its values to enhance the layout of our website pages.