Loops in JavaScript
Today I learn about Loop and its types.
Loops are use for repeated statement executions. It has major 3 types. for, while and do-while loop but for loop further 2 parts divide one is for-in and second is for-of. for-in use for objects details and for-of use for array and string details. Loops are very important role in any programming language because it`s work execution any number statement.